refactor: ThemeProvider 적용하여 기획전 부분 스타일 리팩토링
